requirements.txt
Pillow
plotly
requests
streamlit
streamlit_lottie
streamlit_extras
